Create a tabbar screen to demo loading fonts from Bundle and from URL
The previous TableViewController is now BundleVC.
The load from URL is not yet implemented.
