+++ FontBlasterExample/FontBlasterExample/TabVC.swift
@@ -1,0 +1,15 @@
+// ©2019 Mihai Cristian Tanase. All rights reserved.
+
+import UIKit
+
+class TabVC: UITabBarController {
+  override func viewDidLoad() {
+    super.viewDidLoad()
+  }
+
+  @IBAction func didPressB() {
+    if let vc = selectedViewController as? BaseVC {
+      vc.loadFonts()
+    }
+  }
+}
